functions/memory-consolidator/function_app.py
import azure.functions as func

app = func.FunctionApp()

@app.timer_trigger(schedule="0 0 * * * *", arg_name="mytimer", run_on_startup=False)
def memory_consolidator(mytimer: func.TimerRequest) -> None:
    """Memory consolidation function - placeholder for FR-11 implementation."""
    pass